requirements.txt
allure-pytest
pytest
pandas
requests
pyyaml
jsonschema
pytest-html
selenium
